tests: Fix
